Moved notificationHelper.js to api folder
